changed OrderPage -> MakeTable
