+++ src/App.js
@@ -1,11 +1,12 @@
 import './App.css';
 import 'bulma/css/bulma.min.css';
 import OrderPage from './pages/OrderPage';
+import MakeTable from './pages/MakeTable';
 
 function App() {
   return (
     <>
-      <OrderPage />
+      <MakeTable />
     </>
   );
 }
